% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceandf-package.R
\docType{package}
\name{oceandf-package}
\alias{oceandf}
\alias{oceandf-package}
\title{oceandf: Read 'ODF' (Ocean Data File) Files}
\description{
Reads 'ODF' (Ocean Data File) file table information
  as a data frame and header information as a list.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey.dunnington@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
