% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-odf.R
\name{read_odf}
\alias{read_odf}
\alias{read_odf_colmeta}
\alias{read_odf_header}
\alias{read_odf_header_lines}
\alias{odf_guess}
\title{Read 'ODF' files}
\usage{
read_odf(file, skip = 0, n_max = Inf)

read_odf_colmeta(file, header = read_odf_header(file))

read_odf_header(file, header_lines = read_odf_header_lines(file))

read_odf_header_lines(file, n_header = 1000)

odf_guess()
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{skip}{Number of non-header rows to skip.}

\item{n_max}{Maximum number of rows to read.}

\item{header}{A previously read value obtained from \code{\link[=read_odf_header]{read_odf_header()}}.}

\item{header_lines}{A previously read value obtained from
\code{\link[=read_odf_header_lines]{read_odf_header_lines()}}.}

\item{n_header}{The starting guess for number of header lines.}
}
\description{
ODF (Ocean Data) files are whitespace-delimited files
whose column names and are specified in a loosely formatted header.
}
