% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf-example.R
\name{odf_example}
\alias{odf_example}
\title{ODF example files}
\usage{
odf_example(path)
}
\arguments{
\item{path}{An ODF example name}
}
\value{
The full path to the file
}
\description{
ODF example files
}
\examples{
odf_example("CTD_98911_10P_11_DN.ODF")

}
