% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-odf.R
\name{read_odf}
\alias{read_odf}
\alias{odf_guess_col_names}
\alias{odf_guess_col_types}
\title{Read 'ODF' files}
\usage{
read_odf(
  file,
  col_names = NULL,
  col_types = NULL,
  n_max = -1,
  file_encoding = "UTF-8"
)

odf_guess_col_names(
  file,
  parameter_header = read_odf_parameter_header(file, file_encoding = file_encoding),
  file_encoding = "UTF-8"
)

odf_guess_col_types(
  file,
  col_names = odf_guess_col_names(parameter_header = parameter_header),
  parameter_header = read_odf_parameter_header(file, file_encoding = file_encoding),
  file_encoding = "UTF-8"
)
}
\arguments{
\item{file}{A file, URL, or connection. Files ending in .gz, .bz2, .xz, or
.zip will be automatically uncompressed; URLs will be automatically
downloaded. See \code{\link[readr:read_lines]{readr::read_lines()}} for a full description of how
this parameter is interpreted.}

\item{col_names}{A vector of column names or \code{NULL} to guess
using \code{\link[=odf_guess_col_names]{odf_guess_col_names()}}}

\item{col_types}{A \code{\link[readr:cols]{readr::cols()}} spec or \code{NULL} to guess
using \code{\link[=odf_guess_col_types]{odf_guess_col_types()}}}

\item{n_max}{Maximum number of rows to read.
\code{\link[=read_odf_header_lines]{read_odf_header_lines()}}.}

\item{file_encoding}{The encoding used to encode the file. The default
(UTF-8) reflects a guess based on a number of example ODF files.}

\item{parameter_header}{A previously read value obtained from
\code{\link[=read_odf_parameter_header]{read_odf_parameter_header()}}.}
}
\description{
ODF (Ocean Data) files are whitespace-delimited files
whose column names and are specified in a loosely formatted header.
}
\examples{
odf_file <- odf_example("CTD_98911_10P_11_DN.ODF")
read_odf(odf_file)

odf_guess_col_names(odf_file)
odf_guess_col_types(odf_file)

}
